fix: refactoring and cleaning the code
