+++ .idea/libraries/Maven__jakarta_persistence_jakarta_persistence_api_3_1_0.xml
@@ -1,0 +1,13 @@
+<component name="libraryTable">
+  <library name="Maven: jakarta.persistence:jakarta.persistence-api:3.1.0">
+    <CLASSES>
+      <root url="jar://$MAVEN_REPOSITORY$/jakarta/persistence/jakarta.persistence-api/3.1.0/jakarta.persistence-api-3.1.0.jar!/" />
+    </CLASSES>
+    <JAVADOC>
+      <root url="jar://$MAVEN_REPOSITORY$/jakarta/persistence/jakarta.persistence-api/3.1.0/jakarta.persistence-api-3.1.0-javadoc.jar!/" />
+    </JAVADOC>
+    <SOURCES>
+      <root url="jar://$MAVEN_REPOSITORY$/jakarta/persistence/jakarta.persistence-api/3.1.0/jakarta.persistence-api-3.1.0-sources.jar!/" />
+    </SOURCES>
+  </library>
+</component>
